refactor: extract shared parseVerdict helper into src/providers/
parseVerdict + VerdictSchema + tryParseJson were private to
claude-agent-sdk.ts. Three providers (Claude, Codex, Copilot) will
all hand raw validator text to the same fail-closed JSON+shape
parse, so promote the helper to its own module with its own tests.

claude-agent-sdk.ts now imports parseVerdict from
./parse-verdict.js instead of defining it locally. Behavior is
unchanged and covered by both provider.test and the new
parse-verdict.test.
