+++ package.json
@@ -82,6 +82,7 @@
   },
   "dependencies": {
     "@anthropic-ai/claude-agent-sdk": "^0.2.114",
+    "@openai/codex-sdk": "^0.124.0",
     "jiti": "^2.6.1",
     "picomatch": "^4.0.4",
     "zod": "^4.3.6"
